Cabins vector data

Dataset containing the location of tourist private cabins is Southern Norway, within the study area for the oneimpact package. It corresponds to some specific building types (object_type = "Bygning", byggtyp_nbr = c("161", "162", "163")) form the public N50 dataset. The map was cut for the study area presented in the oneimpact package.

Format

A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector file presents the following columns:

cat: Line number, corresponding to the original dataset

byggtyp_nbr: Type of building (code) in the original dataset

kommune: Code of the municipality where the cabin is located

value: Value 1, to be used for rasterization purposes

Source

https://register.geonorge.no/det-offentlige-kartgrunnlaget/n50-kartdata/ea192681-d039-42ec-b1bc-f3ce04c189ac

Examples

(s <- system.file("vector/cabins_vect.gpkg", package = "oneimpact"))
#> [1] "/home/runner/work/_temp/Library/oneimpact/vector/cabins_vect.gpkg"
sf::st_read(s)
#> Reading layer `cabins_vect' from data source 
#>   `/home/runner/work/_temp/Library/oneimpact/vector/cabins_vect.gpkg' 
#>   using driver `GPKG'
#> Simple feature collection with 6875 features and 4 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 146900.1 ymin: 6622822 xmax: 194694.6 ymax: 6658891
#> Projected CRS: ETRS89 / UTM zone 33N
# or
terra::vect(s)
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 6875, 4  (geometries, attributes)
#>  extent      : 146900.1, 194694.6, 6622822, 6658891  (xmin, xmax, ymin, ymax)
#>  source      : cabins_vect.gpkg
#>  coord. ref. : ETRS89 / UTM zone 33N (EPSG:25833) 
#>  names       :    cat byggtyp_nbr kommune value
#>  type        :  <int>       <chr>   <int> <int>
#>  values      : 131621         161     604     1
#>                131623         161     604     1
#>                131627         161     604     1