Calculates the zone of influence from the nearest feature and the cumulative zone of influence of multiple features

This function takes in a raster with locations of infrastructure and calculates (1) a raster representing the distance from each pixel to the neareast feature and (2) a raster (or set of rasters, in case there is more the one value for scale) representing the density of features in space (through a spatial filter/neighborhood analysis). The neighborhood analysis is done with the raster::focal() function.

Usage

calc_zoi(
  x,
  zoi_radius,
  transform_nearest = NULL,
  type_cumulative = c("circle", "Gauss", "rectangle", "mfilter")[1],
  extent_x_cut = bbox(x)[1, ],
  extent_y_cut = bbox(x)[2, ],
  ...
)

Arguments

x

[RasterLayer,SpatRaster]
Raster representing locations of features, preferentially with value 1 (or any other positive value) where the features are located and NA elsewhere. Alternatively, x might be a binary (dummy) spatial variable representing the presence of linear or area features, with NA/no-data as background. x can be a RasterLayer from raster package or a SpatRaster from terra package. If where = "GRASS", x must be a string corresponding to the name of the input map within a GRASS GIS location and mapset. Maps without NA as background might be prepared as input for calc_zoi_nearest through raster algebra in R and e.g. through the use of the module r.null in GRASS GIS.

zoi_radius

[numeric(1)]
Zone of Influence (ZoI) radius, the distance at which the ZoI vanishes or goes below a given minimum limit value zoi_limit. See zoi_functions() for details. This parameter is ignored if type = "euclidean", type = "log", or type = "sqrt".

extent_x_cut

[numeric vector(2)=c(0,1)]
Vector representing the minimum and maximum extent in x and y for the final output, in the format c(min,max). It is intended to keep only a region of interest, for standardizing the parameters and region when comparing the resulting ZoI maps with the cumulative ZoI, calculated through calc_zoi_cumulative().

...

Adittional parameters passed to terra::distance() or to the ZoI functions (see zoi_functions()) when the calculations are performed in R. No additional parameters implemented for computation in GRASS GIS.

Value

A RasterBrick with de distance to the nearest feature and the densities for all scales selected.

Details

The neighborhood analysis can be done with different methods. The default is a Gaussian filter (type_density = "Gauss"), in which case scale corresponds to the sigma paramater of the Gaussian filter. If type_density = "circle" or type_density = "rectangle", the scale corresponds to the radius of the circle or width of the rectangle, respectively. See raster::focalWeight() for more details.

TO IMPROVE2: do the same in communication with GRASS GIS.

Examples

library(mobsim)
#> Error in library(mobsim): there is no package called ‘mobsim’
library(raster)
#> Loading required package: sp

set.seed(1234)

# set points
ext <- 30000
wd <- ext/20
pts <- set_points(n_features = 1000, centers = 1,
                  width = wd, res = 100,
                  extent_x = c(0, ext), extent_y = c(0, ext),
                  buffer_around = 10000)
#> Error in loadNamespace(x): there is no package called ‘mobsim’
plot(pts$pts)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'pts' not found
plot(pts$rast)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'pts' not found

# calculate distance and densities considering only the initial extent
zoi_values <- c(250, 500, 1000, 2500, 5000)/2
nearest_cumulative <- calc_zoi(pts$rast, zoi = zoi_values,
                                     type_cumulative = "Gauss",
                                     extent_x_cut = c(0, ext), extent_y_cut = c(0, ext))
#> Error in calc_zoi(pts$rast, zoi = zoi_values, type_cumulative = "Gauss",     extent_x_cut = c(0, ext), extent_y_cut = c(0, ext)): could not find function "calc_zoi"
plot(nearest_cumulative)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'nearest_cumulative' not found

# calculate exp_decay (the rest is equal)
log_dist_dens <- calc_zoi(pts$rast, type_density = "Gauss", scale = scales,
                                transform_dist = "exp_decay", half_life = 200,
                                extent_x_cut = c(0, ext), extent_y_cut = c(0, ext))
#> Error in calc_zoi(pts$rast, type_density = "Gauss", scale = scales, transform_dist = "exp_decay",     half_life = 200, extent_x_cut = c(0, ext), extent_y_cut = c(0,         ext)): could not find function "calc_zoi"
plot(log_dist_dens)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'log_dist_dens' not found

# set points
ext <- 30000
wd <- ext/20
pts <- set_points(n_features = 1000, centers = 1,
                  width = wd, res = 100,
                  extent_x = c(0, ext), extent_y = c(0, ext),
                  buffer_around = 0)
#> Error in loadNamespace(x): there is no package called ‘mobsim’
plot(pts$pts)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'pts' not found
plot(pts$rast)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'pts' not found

# calculate distance and densities considering only the initial extent
scales <- c(250, 500, 1000, 2500, 5000)/2
dist_dens <- calc_zoi(pts$rast, type_density = "Gauss", scale = scales, pad = T)
#> Error in calc_zoi(pts$rast, type_density = "Gauss", scale = scales, pad = T): could not find function "calc_zoi"
plot(dist_dens)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'dist_dens' not found