Binarize continuous raster maps

Through GRASS GIS, this function transforms continuous or categorical maps (with more than 1 category) into binary maps (with only two values), to represent, e.g., habitat-matrix maps in the context of landscape ecology. It can also be used to produce binary maps form maps with only one category (and all the rest as NULL/no-data). It requires an active connection between the R session and a GRASS GIS location and mapset (through the package rgrass7), and that the input map is already loaded within this GRASS GIS mapset.

Usage

grass_binarize(
  x,
  breaks = 0.5,
  output = paste0(x, "_bin"),
  null = NULL,
  setnull = NULL,
  bin_values = c(0, 1),
  input_as_region = FALSE,
  quiet = TRUE,
  verbose = FALSE,
  overwrite = FALSE,
  ...
)

Arguments

x

[character(1)]
Name of the input raster map, within a GRASS GIS location and mapset.

breaks

[numeric]
Breaks or threshold to define the binary values in the output binary map. All pixels with value >= breaks are considered as 1 (or the upper value defined in bin_values), and all the rest are considered as 0 (or the lower value defined in bin_values). breaks might be either a single numeric value or a vector of numeric values, in which case multiple binary maps are created (with different break thresholds).

output

[character(1)]
Name of the output map, or prefix of the output map if length(breaks) > 1. In the latter case, the names are completed with the break value. The defult is to use the same name as the input map, plus "_bin" in the end.

null

[numeric(1)=NULL]
If NULL (default), all NULL/no-data pixels in from x are kept as they are in the output map. Otherwise, a numeric value that all NULL pixels should assume in the output map. It uses the module r.null).

setnull

[]
If NULL (default), no changes are made. Otherwise, a set of numeric values that should be transformed into NULL/NA data (using the module r.null).

bin_values

[numeric(2)=c(0,1)]
Values c(lower, upper) that the output map pixels should have if their values are either "lower" or "equal or higher" breaks. By default, c(0, 1).

quiet, overwrite

[logical(1)]
Whether the procedures is GRASS GIS should be run quetly (flag quiet = TRUE) and whether the output maps should be overwriten (flag overwrite = TRUE).

Value

A binarized map with only two values (or a set of binarized maps if length(breaks) > 1).

Details

For a similar procedure within R, use the function landscapetools::util_binarize() or the raster algebra functions within the raster and terra packages.

See also

Examples

# libraries
library(rgrass7)
library(raster)
#> 
#> Attaching package: ‘raster’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
library(terra)
library(dplyr)

# connect to grass gis 7.8
grassdir <- system("grass78 --config path", intern = T)
#> Error in system("grass78 --config path", intern = T): error in running command
gisDB <- "/data/grass"
loc <- "ETRS_33N/"
ms <- "u_bb_cuminf"
rgrass7::initGRASS(gisBase = grassdir,
                   home = tempdir(),
                   override = T,
                   gisDbase = gisDB,
                   location = loc,
                   mapset = ms)
#> Error in file.exists(gisBase): object 'grassdir' not found

# input map (not binarized)
rgrass7::use_sp()
#> Error in rgrass7::use_sp(): sp or rgdal is not available. You need both packages installed.
name_var <- "private_cabins_sub"

# binarize the input map

# map with only 1
cabins_bin1_name <- grass_binarize(name_var, output = "cabins_bin1",
                                   breaks = 1, overwrite = T)
#> Warning: error in running command
#> Error in parseGRASS(cmd, legacyExec = legacyExec): The command
#>    r.mapcalc.simple --interface-description
#> could not be run (127), and produced the error message:
#>    sh: 1: r.mapcalc.simple: not found
# map with 0, 1
cabins_bin2_name <- grass_binarize(name_var, output = "cabins_bin2",
                                   breaks = 1, null = 0, overwrite = T)
#> Warning: error in running command
#> Error in parseGRASS(cmd, legacyExec = legacyExec): The command
#>    g.copy --interface-description
#> could not be run (127), and produced the error message:
#>    sh: 1: g.copy: not found

# visualize
cabins_bin1_2 <- rgrass7::readRAST(c(cabins_bin2_name, cabins_bin1_name)) %>%
  raster::stack() %>%
  terra::rast()
#> Warning: Package rgrass7 is transitioning to package rgrass for GRASS GIS 8.
#> 'readRAST' is deprecated. Use 'read_RAST' instead.
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rast': error in evaluating the argument 'x' in selecting a method for function 'stack': object 'cabins_bin2_name' not found
plot(cabins_bin1_2, main = c("Binarized map setting null to 0", "Binarized map keeping null"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'cabins_bin1_2' not found

# binarize the map with multiple break values

# first create a continuous map
cont_map_name <- calc_influence_nearest(name_var, zoi = 1000, transform = "exp_decay",
                                        where = "GRASS", overwrite = T)
#> Error in calc_influence_nearest(name_var, zoi = 1000, transform = "exp_decay",     where = "GRASS", overwrite = T): could not find function "calc_influence_nearest"
# binarize
cabins_bin2vals_name <- grass_binarize(cont_map_name, output = "cabins_bin",
                                       breaks = c(0.3, 0.5), overwrite = T)
#> Error in doGRASS(cmd, flags = flags, ..., parameters = parameters, echoCmd = echoCmd,     legacyExec = legacyExec): object 'cont_map_name' not found
# visualize
cabins_bin2vals <- rgrass7::readRAST(c(cont_map_name, cabins_bin2vals_name)) %>%
  raster::stack() %>%
  terra::rast()
#> Warning: Package rgrass7 is transitioning to package rgrass for GRASS GIS 8.
#> 'readRAST' is deprecated. Use 'read_RAST' instead.
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rast': error in evaluating the argument 'x' in selecting a method for function 'stack': object 'cont_map_name' not found
plot(cabins_bin2vals,
     main = c("Original map",
              "Binarized map, break = 0.3",
              "Binarized map, break = 0.5"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'cabins_bin2vals' not found