Rasterizes a vector counting the number of features in each pixel

Add within and other arguments for grass functions as options/parameters

Usage

grass_v2rast_count(
  x,
  output = paste0(x, "_count"),
  column = NULL,
  input_as_region = FALSE,
  align = NULL,
  remove_intermediate = TRUE,
  quiet = TRUE,
  verbose = FALSE,
  overwrite = FALSE,
  ...
)

Arguments

x

Input vector map.

output

Output map name.

column

[chracter(1)=NULL
Default is NULL. If not NULL, the name of a column in the input vector x that corresponds to the column to be summed to count the number of features in each pixel of the output raster map. If NULL, this column is created in a temporary vector, with all values equal 1.

input_as_region

[logical(1)=FALSE]
Default is FALSE. Whether the GRASS GIS computational region should be set within the function (to the extent of x) or not. If FALSE, the current computational region is used.

align

[character(1)=NULL]
Name of a raster map with which to align the computational region to produce the output map.

Examples

# libraries
library(rgrass7)
library(raster)
library(terra)
library(dplyr)
library(sf)

# connect to grass gis 7.8
grassdir <- system("grass78 --config path", intern = T)
gisDB <- "/data/grass"
loc <- "ETRS_33N/"
ms <- "u_bb_cuminf"
rgrass7::initGRASS(gisBase = grassdir,
                   home = tempdir(),
                   override = T,
                   gisDbase = gisDB,
                   location = loc,
                   mapset = ms)
#> Warning: cannot create dir '/data/grass', reason 'Permission denied'
#> Warning: cannot create dir '/data/grass/ETRS_33N', reason 'No such file or directory'
#> Warning: cannot create dir '/data/grass/ETRS_33N//PERMANENT', reason 'No such file or directory'
#> Warning: cannot create dir '/data/grass/ETRS_33N//u_bb_cuminf', reason 'No such file or directory'
#> Warning: running command 'g.version' had status 1
#> Error in if (gv == "sh: line 1: g.version: command not found") {    compatible <- as.logical(NA)    attr(compatible, "message") <- gv    return(compatible)}: argument is of length zero

# input region
rgrass7::use_sf()
#> Error in rgrass7::use_sf(): sf or stars is not available. You need both packages installed.
(s <- system.file("vector/study_area.gpkg", package = "oneimpact"))
#> [1] ""
test_region_name <- "region_test_oneimpact_pkg"
sf::st_read(s) %>%
  rgrass7::writeVECT(vname = test_region_name, v.in.ogr_flags = "overwrite")
#> Warning: Package rgrass7 transitioning to package rgrass for GRASS 8.
#> 'writeVECT is deprecated. Use 'write_VECT' instead.
#> Error in rgrass7::writeVECT(., vname = test_region_name, v.in.ogr_flags = "overwrite"): rgdal not available

# set region
rgrass7::execGRASS("g.region", vector = test_region_name, res = "100", flags = c("p"))
#> Error in rgrass7::execGRASS("g.region", vector = test_region_name, res = "100",     flags = c("p")): The command:
#> g.region -p vector=region_test_oneimpact_pkg res=100
#> produced an error (1) during execution:
#> access: No such file or directory
#> ERROR: LOCATION </data/grass/ETRS_33N/> not available

# input vector
(c <- system.file("vector/cabins_vect.gpkg", package = "oneimpact"))
#> [1] ""
cabins_vect_name <- "private_cabins_vect"
sf::st_read(c) %>%
  rgrass7::writeVECT(vname = cabins_vect_name, v.in.ogr_flags = "overwrite")
#> Warning: Package rgrass7 transitioning to package rgrass for GRASS 8.
#> 'writeVECT is deprecated. Use 'write_VECT' instead.
#> Error in rgrass7::writeVECT(., vname = cabins_vect_name, v.in.ogr_flags = "overwrite"): rgdal not available

# rasterize with count, creating a new temp_vector
cabins_count_name <- grass_v2rast_count(cabins_vect_name, output = "cabins_count",
                                        quiet = F, overwrite = T)
#> Error in parseGRASS(cmd, legacyExec = legacyExec): The command
#>    v.in.region --interface-description
#> produced an error (1) during execution:
#>    access: No such file or directoryERROR: LOCATION </data/grass/ETRS_33N/> not available

# rasterize with count, without creating a temporary vector
cabins_count_name <- grass_v2rast_count(cabins_vect_name, output = "cabins_count",
                                        column = "value", quiet = F, overwrite = T)
#> Error in parseGRASS(cmd, legacyExec = legacyExec): The command
#>    v.out.ascii --interface-description
#> produced an error (1) during execution:
#>    access: No such file or directoryERROR: LOCATION </data/grass/ETRS_33N/> not available

# visualize
rgrass7::use_sp()
#> Error in rgrass7::use_sp(): sp or rgdal is not available. You need both packages installed.
rgrass7::readRAST(cabins_count_name) %>%
  raster::raster() %>%
  terra::rast() %>%
  plot(main = "Number of private cabins")
#> Warning: Package rgrass7 is transitioning to package rgrass for GRASS GIS 8.
#> 'readRAST' is deprecated. Use 'read_RAST' instead.
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': error in evaluating the argument 'x' in selecting a method for function 'rast': error in evaluating the argument 'x' in selecting a method for function 'raster': object 'cabins_count_name' not found

# remove rasters created
to_remove_vect <- c(test_region_name, cabins_vect_name)
to_remove_rast <- c(cabins_count_name)
#> Error in eval(expr, envir, enclos): object 'cabins_count_name' not found
# rgrass7::execGRASS("g.remove", type = "vect", name = to_remove_vect, flags = "f")
# rgrass7::execGRASS("g.remove", type = "rast", name = to_remove_rast, flags = "f")