GPS positions for wild reindeer in Norway.

A dataset containing GPS positions for wild reindeer (Rangifer tarandus tarandus) in Setesdal Austhei management area, southern Norway. It includes data from 9 female individuals collected between 2007 and 2010. The coordinates are presented in coordinate reference system ETRS89/UTM 33N (EPSG:25833).

Dataset taken from the supplementary data from the paper Cagnacci, F., Focardi, S., Ghisla, A., van Moorter, B., Merrill, E.H., Gurarie, E., Heurich, M., Mysterud, A., Linnell, J., Panzacchi, M., May, R., Nygård, T., Rolandsen, C. and Hebblewhite, M. (2016), How many routes lead to migration? Comparison of methods to assess and characterize migratory movements. J Anim Ecol, 85: 54-68. https://doi.org/10.1111/1365-2656.12449.

Usage

reindeer

Format

A data frame with 33,202 rows and 6 variables:

x

GPS relocations expressed as UTM easting coordinates, in ETRS89/UTM 33N

y

GPS relocations expressed as UTM northing coordinates, in ETRS89/UTM 33N

t

Timestamp of GPS relocations, in UTC

original_animal_id

Animal name assigned at captures

animal_year_id

Unique individual identifier in the population, and sample unit- i.e. data from one animal in one year

sex

Sex of the individual animal

Source

Panzacchi, M., Van Moorter, B., Strand, O., Loe, L. E., & Reimers, E. (2015). Searching for the fundamental niche using individual-based habitat selection modelling across populations. Ecography, 38(7), 659–669. https://doi.org/10.1111/ecog.01075

Cagnacci, F., Focardi, S., Ghisla, A., van Moorter, B., Merrill, E.H., Gurarie, E., Heurich, M., Mysterud, A., Linnell, J., Panzacchi, M., May, R., Nygård, T., Rolandsen, C. and Hebblewhite, M. (2016), How many routes lead to migration? Comparison of methods to assess and characterize migratory movements. J Anim Ecol, 85: 54-68. https://doi.org/10.1111/1365-2656.12449.

Dataset: https://datadryad.org/stash/dataset/doi:10.5061/dryad.rg0v3

Examples

library(tibble)
data("reindeer")