Simulate landscapes with features and calculate distance and densities of features

This function is a combination of set_points() and calc_dist_dens(), to simulate a point pattern representing the distribution of point-type infrastructure in space, rasterize it, and calculate the distance to and density of infrastructure.

Usage

simulate_dist_dens(
  n_features = 1000,
  centers = 1,
  width = 0.05,
  res = 0.1,
  extent_x = c(0, 1),
  extent_y = c(0, 1),
  buffer_around = 0,
  transform_dist = NULL,
  log_base = exp(1),
  dist_offset = 1,
  type_density = c("Gauss", "circle", "rectangle")[1],
  scale = 100,
  extent_x_cut = extent_x,
  extent_y_cut = extent_y
)

Arguments

n_features

[integer(1)=1000]
Total number of features to spread in space.

centers

[integer(1)=1]
Number of centers around which the features will be placed. Used only if method = "mobsim".

width

[numeric(1)=0.05]
Mean distance between each of the features in a cluster and the center of the cluster. Used only if method = "mobsim".

res

[numeric(1)=0.1]
Resolution of the output raster.

extent_x

[numeric vector(2)=c(0,1)]
Vector representing the minimum and maximum extent in x and y within which the points should be placed, in the format c(min,max).

buffer_around

[numeric(1)=0.1]
Size of the buffer around the extent of the landscape, to avoid edge effects when calculating densities using neighborhood analysis.

Value

A RasterBrick with de distance to the nearest feature and the densities for all scales selected.

Details

The function builds upon the function mobsim::sim_thomas_community() from the mobsim package. Originally the function is intended to simulate positions of multiple species in the context of species abundance distribution studies, but it fits well in case of a single species (or point patterns for a single type of feature).

TO IMPROVE: implement rasterization with terra package

Examples

library(mobsim)
#> Error in library(mobsim): there is no package called ‘mobsim’
library(raster)

set.seed(1234)

# multiple focii
ext <- 30000
wd <- 0.1*ext
scales <- c(250, 500, 1000, 2500, 5000)/2
rasts <- simulate_dist_dens(n_features = 1000, centers = 5,
                            width = wd, res = 100,
                            extent_x = c(0, ext), extent_y = c(0, ext),
                            buffer_around = 10000,
                            type_density = "Gauss", scale = scales,
                            extent_x_cut = c(0, ext), extent_y_cut = c(0, ext))
#> Error in simulate_dist_dens(n_features = 1000, centers = 5, width = wd,     res = 100, extent_x = c(0, ext), extent_y = c(0, ext), buffer_around = 10000,     type_density = "Gauss", scale = scales, extent_x_cut = c(0,         ext), extent_y_cut = c(0, ext)): could not find function "simulate_dist_dens"
plot(rasts)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'rasts' not found