Calculates the zone of influence from the nearest feature and the cumulative zone of influence of multiple features

This function takes in a raster with locations of infrastructure and calculates (1) a raster representing the distance from each pixel to the neareast feature and (2) a raster (or set of rasters, in case there is more the one value for scale) representing the density of features in space (through a spatial filter/neighborhood analysis). The neighborhood analysis is done with the raster::focal() function.

Usage

calc_zoi(
  x,
  zoi_metric = c("all", "nearest", "cumulative")[1],
  where = c("R", "GRASS")[1],
  ...
)

Arguments

x

[RasterLayer,SpatRaster]
Raster representing locations of features, preferentially with value 1 (or any other positive value) where the features are located and NA elsewhere. Alternatively, x might be a binary (dummy) spatial variable representing the presence of linear or area features, with NA/no-data as background. x can be a RasterLayer from raster package or a SpatRaster from terra package. If where = "GRASS", x must be a string corresponding to the name of the input map within a GRASS GIS location and mapset. Maps without NA as background might be prepared as input for calc_zoi_nearest through raster algebra in R and e.g. through the use of the module r.null in GRASS GIS.

zoi_metric

[character(1)="all"]{"all", "nearest", "cumulative"}
Which metric of zone of influence should be computed. Either "all", "nearest", or "cumulative".

where

[character(1)="R"]{"R", "GRASS"}
Where should the computation be done? Default is "R". If where = "GRASS", the R session must be linked to an open GRASS GIS session in a specific location and mapset.

...

Adittional parameters passed to terra::distance() or to the ZoI functions (see zoi_functions()) when the calculations are performed in R. No additional parameters implemented for computation in GRASS GIS.

Value

A RasterBrick with de distance to the nearest feature and the densities for all scales selected.

Details

The neighborhood analysis can be done with different methods. The default is a Gaussian filter (type_density = "Gauss"), in which case scale corresponds to the sigma paramater of the Gaussian filter. If type_density = "circle" or type_density = "rectangle", the scale corresponds to the radius of the circle or width of the rectangle, respectively. See raster::focalWeight() for more details.

TO IMPROVE2: do the same in communication with GRASS GIS.

See also

Examples

library(terra)
#> terra 1.6.7