Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.044363662 0.086524732 0.043310914 0.020290784 0.016665253 0.073244308
#>   [7] 0.023438548 0.035755345 0.021460248 0.016433297 0.067457162 0.005984589
#>  [13] 0.050525908 0.059098323 0.062885672 0.056915093 0.079004127 0.037007268
#>  [19] 0.038396872 0.043961391 0.062491011 0.047545330 0.033705483 0.041599617
#>  [25] 0.013502477 0.015322765 0.034657272 0.053125740 0.072077923 0.016862371
#>  [31] 0.035352343 0.060664004 0.077966319 0.055408905 0.007223014 0.039133745
#>  [37] 0.004533636 0.037324395 0.032870718 0.042566972 0.016160428 0.053604996
#>  [43] 0.019076173 0.022047021 0.000144411 0.043388879 0.114510087 0.058807165
#>  [49] 0.068670228 0.039580327 0.024366468 0.037152536 0.027195803 0.018789979
#>  [55] 0.070196707 0.072792287 0.034214631 0.039276524 0.034498330 0.052391760
#>  [61] 0.076359610 0.035260610 0.022342533 0.107291970 0.048984008 0.097057513
#>  [67] 0.015546955 0.034241730 0.059795004 0.108514023 0.048331821 0.054160056
#>  [73] 0.008606539 0.075744576 0.071256089 0.046055588 0.061386057 0.062097813
#>  [79] 0.075972182 0.017486473 0.047645133 0.063087333 0.070327111 0.032597813
#>  [85] 0.019467538 0.064974741 0.028471357 0.059922300 0.033800095 0.060404753
#>  [91] 0.060226999 0.063733202 0.049613336 0.082981762 0.021583371 0.026424449
#>  [97] 0.106301804 0.145352804 0.038735882 0.084733226 0.062290385
mean_isolation(pts)
#> [1] 0.0484057