Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.021475974 0.072338713 0.132878009 0.173608012 0.087616824 0.025243720
#>   [7] 0.034030932 0.058027867 0.023269586 0.113916315 0.076800649 0.051014725
#>  [13] 0.035282578 0.034173224 0.047077531 0.063914266 0.027021436 0.043891370
#>  [19] 0.061209603 0.040626229 0.016523869 0.009312904 0.043223631 0.058676380
#>  [25] 0.011062195 0.097083940 0.124341739 0.119274876 0.028875838 0.070164831
#>  [31] 0.085555241 0.051062421 0.038640087 0.011176838 0.029759024 0.036966861
#>  [37] 0.070908889 0.067905512 0.037371199 0.084605276 0.102968938 0.054070611
#>  [43] 0.056821642 0.030053664 0.032078408 0.054044333 0.056788152 0.054177061
#>  [49] 0.033360649 0.140984513 0.043097386 0.040652693 0.047590790 0.034787811
#>  [55] 0.088813350 0.031478895 0.053686131 0.042297915 0.014806803 0.103053921
#>  [61] 0.064488013 0.038983259 0.058688627 0.063611173 0.075930255 0.052487700
#>  [67] 0.122602271 0.058996026 0.032726425 0.022347925 0.078039638 0.035955965
#>  [73] 0.078436033 0.040411074 0.111174838 0.023871361 0.085564843 0.027131522
#>  [79] 0.025577114 0.027778093 0.078357969 0.087904163 0.041654744 0.002129992
#>  [85] 0.033596447 0.113517059 0.073283536 0.033803029 0.097807263 0.096207672
#>  [91] 0.056164330 0.075159378 0.092043860 0.051705296 0.095651797 0.064623508
#>  [97] 0.041919239 0.063302295 0.057843982 0.042085721 0.008697561
mean_isolation(pts)
#> [1] 0.05261186