Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.049792650 0.029442046 0.054770983 0.024369275 0.081942978 0.065713487
#>   [7] 0.020254213 0.137229833 0.019257660 0.026854362 0.040877269 0.082890627
#>  [13] 0.116042862 0.040763207 0.027062859 0.122668345 0.014933054 0.059807131
#>  [19] 0.031075429 0.033100657 0.053936258 0.088473856 0.041125028 0.032067418
#>  [25] 0.019216303 0.031967058 0.042184514 0.056810864 0.082250530 0.072426305
#>  [31] 0.020996883 0.013020734 0.057300879 0.034147994 0.069247108 0.013147489
#>  [37] 0.039674581 0.022138760 0.085699652 0.059114194 0.069877575 0.019018181
#>  [43] 0.005012599 0.061070204 0.055749467 0.045612412 0.059274735 0.054954075
#>  [49] 0.040751336 0.048197517 0.062167836 0.120443101 0.043057625 0.075091538
#>  [55] 0.028800276 0.023540829 0.010588898 0.041877426 0.085007741 0.057675027
#>  [61] 0.036455140 0.086096237 0.083157578 0.066106498 0.034683914 0.044930423
#>  [67] 0.057257445 0.086296471 0.082693792 0.076489329 0.042740544 0.025070615
#>  [73] 0.199914930 0.063872890 0.068643977 0.033989500 0.048171961 0.074530719
#>  [79] 0.039438649 0.015984385 0.004699565 0.013385214 0.016931980 0.036364472
#>  [85] 0.022354494 0.010440978 0.051027412 0.070961948 0.082120414 0.017808813
#>  [91] 0.045970942 0.051937770 0.024388748 0.062108051 0.038301245 0.055029830
#>  [97] 0.020189410 0.055332516 0.032090078 0.065557661 0.071434861
mean_isolation(pts)
#> [1] 0.057666