Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.034529679 0.031230011 0.070910091 0.075318515 0.107091136 0.049730084
#>   [7] 0.031660300 0.055469520 0.055327193 0.050179649 0.019364986 0.002215372
#>  [13] 0.050864472 0.114458334 0.046992310 0.074081283 0.029643383 0.033251948
#>  [19] 0.033490374 0.098015021 0.046980759 0.013983430 0.046101932 0.078352648
#>  [25] 0.050044511 0.029083974 0.030395226 0.048565882 0.060223866 0.023465463
#>  [31] 0.007777246 0.015733295 0.036437873 0.060875303 0.049035305 0.031170386
#>  [37] 0.033876429 0.022877317 0.061865373 0.068952286 0.031246204 0.067684973
#>  [43] 0.039930631 0.070653793 0.054801160 0.033037397 0.109810691 0.051676193
#>  [49] 0.040349384 0.011063178 0.038304845 0.032280665 0.069321381 0.057144199
#>  [55] 0.092122602 0.027327541 0.024346129 0.050443143 0.031454977 0.048556983
#>  [61] 0.040340471 0.062041970 0.071717978 0.096612894 0.030052145 0.063759740
#>  [67] 0.041454518 0.039148648 0.097113632 0.026769442 0.040929764 0.077848290
#>  [73] 0.056513112 0.074132106 0.047482016 0.012512488 0.098476242 0.057616167
#>  [79] 0.033637756 0.023474827 0.041220993 0.025017814 0.064638599 0.020063645
#>  [85] 0.020700297 0.023469055 0.045777390 0.059071341 0.023833396 0.102895763
#>  [91] 0.018057937 0.018651147 0.037272568 0.030740983 0.023299567 0.069000263
#>  [97] 0.020125754 0.042832239 0.073792754 0.046050832 0.049050301
mean_isolation(pts)
#> [1] 0.05221047