Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.062904187 0.017140797 0.061001961 0.036520794 0.045370851 0.050130133
#>   [7] 0.085996466 0.089328935 0.065548653 0.083895459 0.060677133 0.056699467
#>  [13] 0.071865077 0.066069477 0.047087413 0.072326960 0.036993944 0.065368596
#>  [19] 0.010112656 0.066370003 0.022247178 0.067450546 0.060524822 0.098074730
#>  [25] 0.054880966 0.104006096 0.042256790 0.064175656 0.057378485 0.053731523
#>  [31] 0.060970091 0.071072808 0.083263270 0.024089874 0.082863994 0.020961794
#>  [37] 0.052741423 0.067632431 0.096437539 0.023761514 0.115412005 0.067663503
#>  [43] 0.009293784 0.020084719 0.022592158 0.007762251 0.051834756 0.022475790
#>  [49] 0.079686337 0.043076851 0.039691460 0.034981843 0.064597692 0.010108542
#>  [55] 0.026675677 0.101786511 0.046443946 0.061410342 0.061476114 0.053551286
#>  [61] 0.052238076 0.082849153 0.069141256 0.044971724 0.056510373 0.046518598
#>  [67] 0.011173821 0.046974830 0.036415276 0.024033270 0.024309738 0.014206642
#>  [73] 0.044414329 0.057600859 0.099827428 0.039371852 0.080571000 0.015589244
#>  [79] 0.056848799 0.035275254 0.056665840 0.030352415 0.030076423 0.053273045
#>  [85] 0.029949216 0.032656501 0.084181296 0.034458622 0.077025874 0.100419839
#>  [91] 0.031613574 0.012216083 0.056697998 0.020824061 0.053623687 0.043166010
#>  [97] 0.039748817 0.063827198 0.042294419 0.086107272 0.030113954
mean_isolation(pts)
#> [1] 0.04802646