Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.076640563 0.099258962 0.023461468 0.017800855 0.078725266 0.033606772
#>   [7] 0.069285921 0.037184088 0.033669023 0.058527990 0.079929701 0.101670203
#>  [13] 0.047005733 0.064762014 0.016527611 0.106495656 0.045827219 0.017265715
#>  [19] 0.016562908 0.004443619 0.069888698 0.100588063 0.063286062 0.013478249
#>  [25] 0.023812838 0.040892467 0.045044208 0.023438551 0.038186641 0.057254286
#>  [31] 0.065830396 0.050144591 0.024254322 0.011699433 0.053472196 0.019426472
#>  [37] 0.038365262 0.013725909 0.076799210 0.019368798 0.068595418 0.009346226
#>  [43] 0.088105378 0.059017366 0.058178898 0.030258648 0.034340172 0.052936431
#>  [49] 0.038091052 0.057468620 0.050638850 0.093388068 0.056614661 0.059729341
#>  [55] 0.035806564 0.066879870 0.041027019 0.045147146 0.055524344 0.067375197
#>  [61] 0.019050046 0.080712769 0.055369513 0.011661932 0.014359656 0.069620053
#>  [67] 0.052357482 0.089334614 0.018803356 0.054095622 0.044080691 0.024542662
#>  [73] 0.077407218 0.037539268 0.059732433 0.093270732 0.029575206 0.043225719
#>  [79] 0.025790657 0.066088115 0.017287023 0.076591992 0.068485418 0.094459929
#>  [85] 0.074571913 0.014855993 0.034301325 0.059505659 0.025703226 0.045869646
#>  [91] 0.050162111 0.019316595 0.037954109 0.046226098 0.039906358 0.053540023
#>  [97] 0.063605047 0.051305586 0.063688027 0.037669098 0.040040654
mean_isolation(pts)
#> [1] 0.05170803