Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.056794478 0.061252406 0.066589498 0.020062225 0.031603868 0.019716710
#>   [7] 0.027524926 0.068103831 0.057479044 0.010251356 0.085775597 0.072311409
#>  [13] 0.060357591 0.016636693 0.086515245 0.022977096 0.056829236 0.023981382
#>  [19] 0.018662741 0.071605539 0.028648597 0.043374063 0.046864940 0.066136698
#>  [25] 0.027738352 0.066123489 0.036925452 0.013361923 0.089476556 0.031373683
#>  [31] 0.095252957 0.045427854 0.050751297 0.078290713 0.025089742 0.031185275
#>  [37] 0.053655817 0.038307322 0.030952492 0.007158434 0.042626801 0.043226941
#>  [43] 0.079379827 0.043608298 0.045571856 0.026532365 0.018782600 0.081820643
#>  [49] 0.034470994 0.036413131 0.028134560 0.062212546 0.090772643 0.055090339
#>  [55] 0.037196087 0.035834307 0.063366094 0.085272852 0.032744222 0.062876753
#>  [61] 0.033087521 0.081101404 0.070408948 0.008966181 0.052598217 0.065024543
#>  [67] 0.031599492 0.146832860 0.018249269 0.065446283 0.109047159 0.030764794
#>  [73] 0.074769330 0.069245728 0.072451719 0.033391326 0.065516298 0.050324879
#>  [79] 0.076184221 0.054855112 0.025948021 0.049167049 0.021936912 0.035242998
#>  [85] 0.043977889 0.038283530 0.054836751 0.023595684 0.053188729 0.052793194
#>  [91] 0.045249231 0.018662722 0.036618275 0.023234610 0.019485638 0.009993030
#>  [97] 0.055223886 0.038920006 0.067116282 0.083275408 0.039558650
mean_isolation(pts)
#> [1] 0.05114849