Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.064542744 0.017302090 0.089667119 0.031489366 0.058250262 0.063648040
#>   [7] 0.045085807 0.033830605 0.058300951 0.046530735 0.065901650 0.039209656
#>  [13] 0.092481037 0.044007378 0.053301502 0.038549317 0.064488399 0.011460276
#>  [19] 0.069850700 0.029264469 0.056826327 0.005529616 0.057933059 0.014748518
#>  [25] 0.007484056 0.028399383 0.031193347 0.037176520 0.020892616 0.052533117
#>  [31] 0.030359353 0.086965231 0.037507196 0.017279991 0.044170621 0.110096954
#>  [37] 0.021800600 0.071464433 0.023049467 0.020418287 0.022763115 0.014155686
#>  [43] 0.020501980 0.065792997 0.079780931 0.007243537 0.098909379 0.075720358
#>  [49] 0.046774719 0.041015354 0.032735484 0.038998473 0.024149783 0.018918035
#>  [55] 0.080308560 0.072730376 0.043078922 0.040062565 0.031756589 0.059930687
#>  [61] 0.046079020 0.017068524 0.047544928 0.045024575 0.025763924 0.049334108
#>  [67] 0.011073566 0.030177061 0.040206570 0.071888552 0.026073747 0.065403551
#>  [73] 0.050446279 0.095071710 0.054506425 0.075470014 0.055923864 0.065620698
#>  [79] 0.031943655 0.085459394 0.010584358 0.042285990 0.048470731 0.063317728
#>  [85] 0.016669306 0.051766551 0.026647757 0.097501008 0.058770285 0.085805444
#>  [91] 0.046816157 0.070529847 0.055757214 0.037047959 0.065438179 0.040752636
#>  [97] 0.049152341 0.036304081 0.058838912 0.055722978 0.031230786
mean_isolation(pts)
#> [1] 0.05129272