Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.017666950 0.009092797 0.078068851 0.021604662 0.047562030 0.041876497
#>   [7] 0.010556336 0.046137334 0.033261530 0.048958333 0.032230764 0.017242640
#>  [13] 0.026685610 0.022648426 0.008349746 0.092816654 0.041093925 0.087713532
#>  [19] 0.146771108 0.060128585 0.054317943 0.037917252 0.085144654 0.074076404
#>  [25] 0.061891088 0.025468982 0.097488658 0.062055065 0.065189133 0.050065449
#>  [31] 0.111896616 0.032188257 0.028201728 0.081464273 0.007533156 0.019153153
#>  [37] 0.080762720 0.010151377 0.028875860 0.099692920 0.033801212 0.020630361
#>  [43] 0.054326865 0.068740083 0.062712455 0.043170445 0.039463118 0.093214945
#>  [49] 0.044239436 0.033300232 0.054272935 0.014141035 0.068997572 0.044618838
#>  [55] 0.077809260 0.022730697 0.020157065 0.063675392 0.048380380 0.019857183
#>  [61] 0.046770378 0.070229728 0.070760859 0.041873872 0.108073768 0.048346962
#>  [67] 0.024527018 0.055274048 0.015738175 0.076875872 0.029874172 0.009053506
#>  [73] 0.064794325 0.060490169 0.056753492 0.104405905 0.033779053 0.062044506
#>  [79] 0.062297962 0.059960886 0.032283038 0.066352700 0.053104254 0.055045342
#>  [85] 0.007565884 0.038760945 0.049280822 0.044115697 0.061212972 0.035289978
#>  [91] 0.088237142 0.072079362 0.057909059 0.084081283 0.009192622 0.089909197
#>  [97] 0.134808202 0.068722732 0.008679640 0.043120955 0.027733662
mean_isolation(pts)
#> [1] 0.05362955