Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.020047087 0.065697888 0.042293952 0.019979062 0.018724349 0.022091191
#>   [7] 0.115318680 0.035297819 0.072040545 0.079685647 0.018957239 0.029753948
#>  [13] 0.029856089 0.040313514 0.032408887 0.038484675 0.022308386 0.027948194
#>  [19] 0.036320426 0.031971405 0.058820183 0.052437141 0.047913152 0.074138073
#>  [25] 0.021830259 0.016824275 0.041443549 0.057916229 0.014730374 0.066405419
#>  [31] 0.092581822 0.018121677 0.015344928 0.124984724 0.019593504 0.079588986
#>  [37] 0.004439786 0.032743652 0.045746241 0.056711654 0.040124616 0.025131494
#>  [43] 0.032023625 0.116007547 0.024586611 0.055005984 0.037687069 0.057738843
#>  [49] 0.073413143 0.100776086 0.108620606 0.089565802 0.077850881 0.069556090
#>  [55] 0.013635853 0.084149153 0.035080074 0.051628482 0.043634169 0.058022801
#>  [61] 0.032308646 0.042417877 0.041689479 0.090112366 0.031273947 0.053599668
#>  [67] 0.060788477 0.044796425 0.060136628 0.059758431 0.048201437 0.022439225
#>  [73] 0.068662325 0.056284777 0.031507389 0.015547003 0.078105206 0.077814961
#>  [79] 0.039068037 0.077219045 0.065828238 0.001878757 0.060500578 0.064059299
#>  [85] 0.046627854 0.072156685 0.092425087 0.069319046 0.047979888 0.023186713
#>  [91] 0.056614380 0.048837892 0.057235662 0.046446505 0.077412110 0.025754691
#>  [97] 0.013989827 0.083527085 0.149055779 0.065584908 0.039909163
mean_isolation(pts)
#> [1] 0.05245114