Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.037905901 0.033155509 0.036259521 0.081576606 0.098690490 0.040358177
#>   [7] 0.058462450 0.022509129 0.040833291 0.054629648 0.063886273 0.051419931
#>  [13] 0.026427877 0.019588937 0.071601496 0.018470280 0.055385199 0.050447271
#>  [19] 0.047250732 0.053183449 0.072387287 0.149109054 0.025512826 0.060445365
#>  [25] 0.036368560 0.097979423 0.010947166 0.046498653 0.064914836 0.075519530
#>  [31] 0.009736066 0.006004334 0.077098150 0.056209963 0.038003124 0.076303640
#>  [37] 0.047942505 0.018786458 0.051897982 0.024262838 0.036952652 0.005152142
#>  [43] 0.060025111 0.041461665 0.047129880 0.068055950 0.037024672 0.093284352
#>  [49] 0.029153498 0.079227523 0.048108898 0.046220358 0.016669020 0.040770669
#>  [55] 0.030768384 0.048662471 0.043527502 0.036098268 0.021867641 0.043158928
#>  [61] 0.031434533 0.048660816 0.034506334 0.107554218 0.022940242 0.037566119
#>  [67] 0.035466276 0.039945949 0.016127726 0.038533000 0.033307440 0.027150023
#>  [73] 0.052061649 0.022768440 0.089547662 0.039789298 0.073202307 0.105802895
#>  [79] 0.066408076 0.052710714 0.088766921 0.054361043 0.059959629 0.025610012
#>  [85] 0.027628865 0.027028153 0.037435617 0.088463772 0.023673105 0.032081854
#>  [91] 0.055784943 0.054894922 0.024872787 0.025528505 0.151061616 0.060890058
#>  [97] 0.046285646 0.008490896 0.015899240 0.089375044 0.127202098
mean_isolation(pts)
#> [1] 0.05380444