Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.078036175 0.047586541 0.041821881 0.076757805 0.045996372 0.019244185
#>   [7] 0.033855454 0.018573149 0.059561178 0.043618974 0.008424903 0.061468162
#>  [13] 0.030367163 0.037346422 0.030185941 0.072795235 0.083844645 0.020043349
#>  [19] 0.031070936 0.040889089 0.087168484 0.027501537 0.058921104 0.026325952
#>  [25] 0.073137236 0.042460581 0.006210345 0.070481908 0.017695640 0.043257563
#>  [31] 0.059475716 0.092211694 0.102129633 0.044414338 0.039089336 0.052562170
#>  [37] 0.032056313 0.005207517 0.077946161 0.046200369 0.009232798 0.092799980
#>  [43] 0.004236877 0.049232956 0.056363563 0.020289469 0.040748053 0.015552964
#>  [49] 0.025256200 0.036607767 0.042901744 0.045659024 0.046460942 0.043320225
#>  [55] 0.134570265 0.090107010 0.047065583 0.069032035 0.035950288 0.149189013
#>  [61] 0.039681046 0.028466085 0.024967692 0.025847316 0.054396845 0.038491033
#>  [67] 0.049040196 0.057964991 0.054400466 0.033598442 0.021640308 0.007588657
#>  [73] 0.091987952 0.042147033 0.041536696 0.054253478 0.070032751 0.114271229
#>  [79] 0.014788866 0.022704108 0.086789841 0.043082562 0.044457463 0.029819506
#>  [85] 0.039832091 0.030610955 0.022473966 0.080199493 0.090638600 0.055875426
#>  [91] 0.060581151 0.037421805 0.141893390 0.058735329 0.066609353 0.032034215
#>  [97] 0.057185508 0.118720811 0.051620285 0.060141177 0.051013059
mean_isolation(pts)
#> [1] 0.05628079