Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.030872999 0.095156887 0.030366026 0.117521067 0.061639326 0.096601212
#>   [7] 0.055780460 0.037238952 0.031136268 0.085737839 0.118755214 0.033328997
#>  [13] 0.042607543 0.028564315 0.063386554 0.084614297 0.065398268 0.103520717
#>  [19] 0.027020821 0.054300570 0.124356854 0.101319003 0.046963097 0.058062083
#>  [25] 0.008303331 0.152971775 0.123176454 0.067792280 0.024676220 0.009700032
#>  [31] 0.045076540 0.091078209 0.024867481 0.053088805 0.096646889 0.022427909
#>  [37] 0.077602311 0.028180435 0.044923026 0.035225424 0.020868347 0.018217310
#>  [43] 0.041305834 0.065918771 0.089409508 0.031487076 0.021571424 0.058314840
#>  [49] 0.032264236 0.067400340 0.025415780 0.058224546 0.031159813 0.039605410
#>  [55] 0.098760513 0.079891769 0.037746402 0.051159175 0.071257040 0.012634436
#>  [61] 0.050565021 0.052158674 0.022049849 0.185448380 0.026015107 0.064398195
#>  [67] 0.059559055 0.061967417 0.029965100 0.098883961 0.029826346 0.055154156
#>  [73] 0.155208659 0.042817622 0.037785059 0.064197444 0.030500019 0.038745368
#>  [79] 0.081462899 0.121605692 0.045612464 0.114248810 0.049485194 0.117891618
#>  [85] 0.071839472 0.028034504 0.126343823 0.059348522 0.094925460 0.042702566
#>  [91] 0.030544609 0.064817231 0.073104723 0.068149332 0.012780157 0.041774072
#>  [97] 0.070830884 0.077020731 0.062902866 0.158436549 0.052452273
mean_isolation(pts)
#> [1] 0.05961596