Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.084953122 0.035835181 0.008389664 0.101741685 0.034431310 0.033548424
#>   [7] 0.013090952 0.062687051 0.065102889 0.019998775 0.040139722 0.057952242
#>  [13] 0.015670010 0.068708617 0.024664002 0.028536806 0.106013549 0.069600116
#>  [19] 0.075518763 0.040746222 0.046529363 0.054097961 0.059463092 0.016189287
#>  [25] 0.009645059 0.073435067 0.014840283 0.020591381 0.069785998 0.113927131
#>  [31] 0.062515143 0.064662817 0.053606664 0.042640852 0.166212961 0.052111442
#>  [37] 0.073567361 0.058509616 0.052438091 0.020776884 0.031599380 0.043128423
#>  [43] 0.017294078 0.039962948 0.035334046 0.042275235 0.052062496 0.046288381
#>  [49] 0.007136565 0.035864284 0.049782571 0.022702394 0.058364637 0.084202579
#>  [55] 0.040723661 0.040483868 0.052079995 0.029609090 0.029227143 0.013838024
#>  [61] 0.129231216 0.022379960 0.086533633 0.057115802 0.059407139 0.054519289
#>  [67] 0.042417862 0.047305226 0.082073298 0.005686371 0.028133728 0.033420103
#>  [73] 0.054678337 0.041986180 0.052964821 0.044481486 0.048969158 0.117086099
#>  [79] 0.061762538 0.056331024 0.015355587 0.050487685 0.044157260 0.038644769
#>  [85] 0.043403124 0.092591751 0.045116359 0.054079889 0.035842384 0.012564392
#>  [91] 0.038782189 0.084924392 0.028781834 0.010706294 0.016195622 0.111677337
#>  [97] 0.033434919 0.100873343 0.034774930 0.014924753 0.059174213
mean_isolation(pts)
#> [1] 0.05350305