Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.055966601 0.029654380 0.039713316 0.098438150 0.055038375 0.040079494
#>   [7] 0.080502531 0.026468051 0.094436474 0.017060259 0.056752913 0.077533336
#>  [13] 0.067202099 0.023602078 0.062764940 0.080632297 0.022804153 0.053584353
#>  [19] 0.045995135 0.037170740 0.040276163 0.047636322 0.055561288 0.025198146
#>  [25] 0.027429849 0.091050713 0.019017200 0.032738746 0.038947908 0.057849407
#>  [31] 0.057480579 0.061073896 0.009621478 0.065630788 0.048787906 0.078450660
#>  [37] 0.024444948 0.028027545 0.046331955 0.066631983 0.027697354 0.047931143
#>  [43] 0.067569902 0.036186179 0.075379348 0.115609279 0.029334432 0.044362100
#>  [49] 0.039767166 0.070618506 0.012926760 0.065125138 0.060209341 0.081260595
#>  [55] 0.090555778 0.077309092 0.038080748 0.024191782 0.066463183 0.017403139
#>  [61] 0.036605983 0.038116591 0.031744288 0.039981802 0.019202187 0.062302584
#>  [67] 0.075991979 0.014401617 0.029364782 0.020240977 0.101489512 0.055995269
#>  [73] 0.028371008 0.074028267 0.011065680 0.031445923 0.047077415 0.068180957
#>  [79] 0.008898735 0.032701754 0.036166443 0.053471409 0.081096028 0.038510322
#>  [85] 0.029869084 0.039459471 0.035460643 0.059652633 0.021795283 0.047935966
#>  [91] 0.036406113 0.101070547 0.032919292 0.046257658 0.021058467 0.045909667
#>  [97] 0.089023767 0.022470520 0.005239836 0.038061809 0.056026997
mean_isolation(pts)
#> [1] 0.04756752