Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.066485069 0.029714640 0.024573425 0.051333716 0.002446050 0.089406091
#>   [7] 0.063598049 0.047931660 0.085559031 0.034113056 0.043864014 0.069644338
#>  [13] 0.084636824 0.043784745 0.057367546 0.042825987 0.015685759 0.059811816
#>  [19] 0.028943200 0.053531351 0.052266606 0.019404727 0.063027160 0.036185050
#>  [25] 0.022943509 0.057847432 0.066714244 0.033069221 0.043615544 0.033536089
#>  [31] 0.073197401 0.001843724 0.037046606 0.038628207 0.033412909 0.083747470
#>  [37] 0.014103847 0.015012938 0.009936140 0.031768071 0.060690954 0.009658661
#>  [43] 0.051142894 0.026918213 0.018045788 0.104247121 0.077110130 0.057502707
#>  [49] 0.044751205 0.027686456 0.037474935 0.051288264 0.039252352 0.014781490
#>  [55] 0.055581028 0.074881280 0.089170591 0.043516926 0.048739940 0.081071103
#>  [61] 0.038062617 0.031008954 0.078736295 0.085040182 0.068398425 0.058775325
#>  [67] 0.118645200 0.074471737 0.078798419 0.044687821 0.048936338 0.065867180
#>  [73] 0.055357530 0.061989500 0.023191156 0.083009702 0.050066230 0.048588764
#>  [79] 0.053314690 0.044072757 0.055212002 0.017769614 0.072815815 0.053643491
#>  [85] 0.079403557 0.021573356 0.078775286 0.037133723 0.097053085 0.062793255
#>  [91] 0.028912761 0.038348890 0.037267139 0.105042153 0.063211224 0.051974073
#>  [97] 0.026803385 0.033151348 0.097776198 0.057114648 0.029322293
mean_isolation(pts)
#> [1] 0.04961497