Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.068346245 0.082614985 0.058174858 0.081319086 0.084725369 0.003557386
#>   [7] 0.064648763 0.021202851 0.053834044 0.029531420 0.033879227 0.010220870
#>  [13] 0.016228332 0.041959348 0.023929249 0.037355686 0.005631982 0.065699796
#>  [19] 0.029028647 0.026111269 0.035569254 0.025759145 0.068757861 0.042535760
#>  [25] 0.021589793 0.091196855 0.051081656 0.037227706 0.057208288 0.084117433
#>  [31] 0.061377869 0.050967412 0.010845361 0.047156657 0.042641689 0.075199765
#>  [37] 0.024262825 0.018863884 0.009532514 0.016928864 0.020732317 0.051164018
#>  [43] 0.086500614 0.044925689 0.084857044 0.060656043 0.064578039 0.084020264
#>  [49] 0.125731939 0.052997937 0.065961365 0.010502925 0.037012062 0.066091656
#>  [55] 0.035831316 0.059872024 0.041850616 0.063150720 0.029502709 0.089315538
#>  [61] 0.067119525 0.059721790 0.057099522 0.062249806 0.007762248 0.047642202
#>  [67] 0.019330189 0.054723474 0.061159675 0.012392616 0.008340875 0.015367765
#>  [73] 0.015273659 0.047029464 0.033976941 0.017484956 0.031713484 0.030418270
#>  [79] 0.023547088 0.003290354 0.074986792 0.048155446 0.080053944 0.075107117
#>  [85] 0.073839792 0.017840702 0.031622227 0.105977163 0.054306101 0.062912255
#>  [91] 0.005764827 0.040389356 0.087272434 0.025929543 0.039309890 0.037133091
#>  [97] 0.053401900 0.038755540 0.066199279 0.043862414 0.057654788
mean_isolation(pts)
#> [1] 0.04878451