Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.049373072 0.059156452 0.070058413 0.030645041 0.108390419 0.056464397
#>   [7] 0.086972580 0.067902624 0.082384047 0.023310840 0.019601063 0.044889406
#>  [13] 0.070042592 0.022554195 0.003280576 0.040322315 0.070038362 0.066668659
#>  [19] 0.064551106 0.049723330 0.021613196 0.015498621 0.055480040 0.046190711
#>  [25] 0.056838109 0.050097308 0.028432664 0.046953352 0.088921804 0.027143419
#>  [31] 0.016583893 0.032005075 0.062757231 0.029404897 0.051908509 0.060882081
#>  [37] 0.053168335 0.030566467 0.026418930 0.046047814 0.059942532 0.026500519
#>  [43] 0.050811952 0.049499710 0.084798354 0.110765633 0.051545870 0.025083895
#>  [49] 0.010711950 0.066211796 0.046934568 0.074628810 0.056315375 0.053871788
#>  [55] 0.031639296 0.075203075 0.058983006 0.112608874 0.050223717 0.071310076
#>  [61] 0.034097528 0.124017851 0.018769846 0.035824015 0.049453465 0.023371931
#>  [67] 0.061889712 0.056868526 0.078818383 0.038445254 0.051053183 0.050053730
#>  [73] 0.035671582 0.100263319 0.029068173 0.006987154 0.050266926 0.055801795
#>  [79] 0.054800491 0.050768241 0.065125057 0.028607257 0.059640277 0.028928267
#>  [85] 0.036831328 0.045808375 0.044445175 0.024301073 0.049728227 0.050770450
#>  [91] 0.065908276 0.076021105 0.099297544 0.049596006 0.014590329 0.021434492
#>  [97] 0.005050301 0.068241373 0.026988081 0.060638756 0.016661163
mean_isolation(pts)
#> [1] 0.05072912