Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.041609068 0.061729802 0.085326445 0.022496098 0.031177594 0.017801902
#>   [7] 0.147490065 0.097220982 0.032288727 0.064230919 0.054909453 0.026333169
#>  [13] 0.016855362 0.072498139 0.079679124 0.053717441 0.048013353 0.043906113
#>  [19] 0.062004693 0.051235708 0.036044817 0.026960638 0.081513367 0.052083667
#>  [25] 0.056103633 0.083174990 0.040252712 0.037370913 0.054372629 0.023378064
#>  [31] 0.079336199 0.067673148 0.023067405 0.044065984 0.085764649 0.123233237
#>  [37] 0.045316662 0.045594647 0.024429806 0.046578438 0.037283774 0.062619123
#>  [43] 0.089663358 0.023532134 0.062712477 0.070848783 0.065401526 0.029061547
#>  [49] 0.053394474 0.017534051 0.153292978 0.045038767 0.009495667 0.009370159
#>  [55] 0.015971613 0.079161127 0.051779711 0.063344810 0.032592529 0.051495611
#>  [61] 0.065332818 0.034601753 0.054064041 0.052491704 0.020143850 0.038179255
#>  [67] 0.078590070 0.034415854 0.119507146 0.056952197 0.114137983 0.084647242
#>  [73] 0.038418689 0.072899149 0.071132414 0.038684274 0.046615071 0.070679867
#>  [79] 0.049511420 0.020896364 0.008358600 0.062060502 0.018719014 0.065806955
#>  [85] 0.065732584 0.036089671 0.022848358 0.052687322 0.026557128 0.096875620
#>  [91] 0.036957792 0.093109282 0.057480702 0.057666354 0.039578717 0.037498982
#>  [97] 0.070826531 0.022827916 0.075707939 0.021154382 0.037089007
mean_isolation(pts)
#> [1] 0.0490172