Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.014236261 0.090995607 0.112333848 0.069238696 0.047682748 0.054776023
#>   [7] 0.086219970 0.043412163 0.085455344 0.028681504 0.028403963 0.126002262
#>  [13] 0.008695082 0.072755106 0.117641062 0.051256806 0.031564846 0.029615858
#>  [19] 0.076612256 0.011349044 0.018892459 0.045042870 0.041568813 0.047377536
#>  [25] 0.036953884 0.093690694 0.073247893 0.003876824 0.056726282 0.087226841
#>  [31] 0.017676594 0.048282196 0.085552941 0.014773658 0.063731051 0.029955266
#>  [37] 0.104691522 0.040551594 0.029201312 0.061662839 0.022822240 0.033898517
#>  [43] 0.003813119 0.006945021 0.042210861 0.047757580 0.093567520 0.014735518
#>  [49] 0.063383125 0.057766539 0.029870840 0.093121614 0.053049471 0.021104001
#>  [55] 0.019866869 0.012464787 0.047067218 0.030074258 0.142649402 0.011783882
#>  [61] 0.037597366 0.098767194 0.018310173 0.052865711 0.098414493 0.039809066
#>  [67] 0.093058330 0.094540912 0.010195261 0.046040005 0.020939766 0.026520806
#>  [73] 0.112211679 0.108142813 0.031877439 0.189878839 0.026355652 0.042035302
#>  [79] 0.034023750 0.128241059 0.045185872 0.034966037 0.064845985 0.055641472
#>  [85] 0.037843818 0.062615000 0.053333531 0.094171711 0.048066616 0.041007139
#>  [91] 0.070019564 0.040056251 0.062146931 0.045357451 0.078663510 0.061828118
#>  [97] 0.013577221 0.070131404 0.024694198 0.058897823 0.025130798
mean_isolation(pts)
#> [1] 0.05683197