Isolation and mean isolation of points in space

Measures of isolation and mean isolation to a set of points in space. isolation() creates random points in a landscape and calculates the nearest neighbor distance from each of them to another set of points passed as input, x. mean_isolation() calculates the average isolation calculated through isolation().

Usage

isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

Arguments

x

[data.frame]
data.frame with (x,y) coordinates in the columns.

n_rand

[numeric(1)=100]
Number of random points to be created in space, to compute the distance to x.

ext

[numeric(x)=c(0,1)] Extent of the space within which the random positions should be created c(x or ymin, x or ymax).

lonlat

[logical(1)=FALSE]
Whether the distance between points should be calculated in an WGS ellipsoid (lonlat = TRUE) or on a plane (lonlat = FALSE). See raster::pointDistance() for more details.

Value

isolation() returns the distance from each random point to the nearest neighbor point in x. mean_isolation() returns the average nearest neighbor distance from all random positions to the points in x.

Details

So far the function only works for a square landscape. In the future we can implement that for polygons or rasters with masks or null cells if necessary, in an approach similar to set_points_sample.

Examples

pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
#>   [1] 0.070223418 0.085821911 0.085737150 0.016000400 0.043994355 0.075585059
#>   [7] 0.045802170 0.104451506 0.016252812 0.026009141 0.088115699 0.093328638
#>  [13] 0.041617149 0.041980175 0.082838365 0.046589518 0.027622517 0.081673882
#>  [19] 0.098349890 0.040480769 0.033339162 0.039278110 0.085972896 0.058239749
#>  [25] 0.085320928 0.039135088 0.016313215 0.050354930 0.090548534 0.069068238
#>  [31] 0.046111986 0.053695391 0.021730333 0.109305672 0.016568806 0.086325477
#>  [37] 0.038748701 0.075906884 0.074921605 0.060953605 0.006754450 0.065971730
#>  [43] 0.047092808 0.055013768 0.125373793 0.020554291 0.095752235 0.086875720
#>  [49] 0.009125345 0.074292920 0.097236320 0.089979210 0.008260726 0.114667116
#>  [55] 0.097168960 0.026926550 0.072004322 0.059707655 0.023364748 0.028498146
#>  [61] 0.040974781 0.047413610 0.067435022 0.030869375 0.024679013 0.022176203
#>  [67] 0.098413208 0.072177446 0.095902647 0.049104885 0.082169651 0.035610884
#>  [73] 0.013006728 0.060324603 0.093050488 0.005631415 0.101081512 0.029176121
#>  [79] 0.033339224 0.097091575 0.079652637 0.082422534 0.021957361 0.068619330
#>  [85] 0.070424167 0.028993100 0.097053565 0.063776645 0.042138559 0.041618263
#>  [91] 0.027873899 0.062909480 0.039618947 0.030830289 0.012057749 0.091894008
#>  [97] 0.061841125 0.048415250 0.034237212 0.082390747 0.027795628
mean_isolation(pts)
#> [1] 0.05365925